// frontend/src/api-generated/models/danhMucThuocTinhResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { LuaChonGoiYResponse } from './luaChonGoiYResponse';

/**
 * Thông tin danh mục thuộc tính trả về
 */
export interface DanhMucThuocTinhResponse {
  /** ID thuộc tính */
  id?: number;
  /** Mã thuộc tính */
  maThuocTinh?: string;
  /** Tên thuộc tính */
  tenThuocTinh?: string;
  /** Kiểu dữ liệu */
  kieuDuLieu?: string;
  /** Áp dụng cho */
  apDungCho?: string;
  /** Bắt buộc nhập */
  batBuocNhap?: boolean;
  /** Giá trị mặc định */
  giaTriMacDinh?: string;
  /** Trạng thái */
  trangThai?: string;
  /** Thời gian tạo */
  thoiGianTao?: string;
  /** Thời gian cập nhật */
  thoiGianCapNhat?: string;
  /** Danh sách tùy chọn gợi ý đi kèm */
  luaChonGoiY?: LuaChonGoiYResponse[];
}