// frontend/src/api-generated/models/danhMucThuocTinhRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { LuaChonGoiYSubRequest } from './luaChonGoiYSubRequest';

/**
 * Yêu cầu tạo/cập nhật danh mục thuộc tính
 */
export interface DanhMucThuocTinhRequest {
  /**
     * Mã thuộc tính duy nhất
     * @minLength 0
     * @maxLength 50
     */
  maThuocTinh?: string;
  /**
     * Tên thuộc tính hiển thị
     * @minLength 0
     * @maxLength 100
     */
  tenThuocTinh: string;
  /**
     * Kiểu dữ liệu (TEXT, NUMBER, SELECT, ...)
     * @minLength 0
     * @maxLength 30
     */
  kieuDuLieu: string;
  /**
     * Phân hệ áp dụng (PHAN_CUNG, PHAN_MEM, LINH_KIEN)
     * @minLength 0
     * @maxLength 50
     * @pattern PHAN_CUNG|PHAN_MEM|LINH_KIEN
     */
  apDungCho: string;
  /** Trường này có bắt buộc nhập hay không */
  batBuocNhap: boolean;
  /**
     * Giá trị mặc định của thuộc tính
     * @minLength 0
     * @maxLength 255
     */
  giaTriMacDinh?: string;
  /**
     * Trạng thái hoạt động (HOAT_DONG/KHOA)
     * @minLength 0
     * @maxLength 30
     */
  trangThai: string;
  /** Danh sách tùy chọn gợi ý đi kèm (Aggregate) */
  luaChonGoiY?: LuaChonGoiYSubRequest[];
}